requirements.txt
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.0
boto3==1.42.42
botocore==1.42.42
cachetools==6.2.6
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.4
deprecation==2.1.0
dj-database-url==3.1.0
Django==5.2.10
django-storages==1.14.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
exceptiongroup==1.3.1
fsspec==2026.1.0
gunicorn==25.0.1
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
jmespath==1.1.0
markdown-it-py==4.0.0
mdurl==0.1.2
mmh3==5.2.0
multidict==6.7.1
packaging==26.0
pillow==12.1.0
postgrest==2.27.3
propcache==0.4.1
psycopg2-binary==2.9.11
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyiceberg==0.10.0
PyJWT==2.11.0
pyparsing==3.3.2
pyroaring==1.0.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
realtime==2.27.3
requests==2.32.5
rich==14.3.2
s3transfer==0.16.0
six==1.17.0
sortedcontainers==2.4.0
sqlparse==0.5.5
storage3==2.27.3
StrEnum==0.4.15
strictyaml==1.7.3
supabase==2.27.3
supabase-auth==2.27.3
supabase-functions==2.27.3
tenacity==9.1.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
websockets==15.0.1
whitenoise==6.11.0
yarl==1.22.0
